hamburg-weather-analytics/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas==2.2.2
duckdb==1.1.3
